Write Lilac data to DB

+++ hmf-common/src/main/java/com/hartwig/hmftools/common/hla/HlaTypeDetails.java
@@ -25,11 +25,11 @@ public interface HlaTypeDetails {
 
     double tumorCopyNumber();
 
-    int somaticMissense();
+    double somaticMissense();
 
-    int somaticNonsenseOrFrameshift();
+    double somaticNonsenseOrFrameshift();
 
-    int somaticSplice();
+    double somaticSplice();
 
-    int somaticSynonymous();
+    double somaticSynonymous();
 }
